Integrated Thomas Fossati's review from 2025-07-17 (RATS mailing list).
https://mailarchive.ietf.org/arch/msg/rats/ktS3-_RDHpN-jITeoqsy0p22oqw

# S 1

* [x] Attestation Policies => Appraisal Policies
  * Renamed.
* [x] It says "The reference models defined can also be applied to the
conveyance of other Conceptual Messages in RATS" without explaining how.
Please remove the sentence: there is already better prose for this in
the last paragraph of this same section.
  * Removed the sentence
* [x] It says: "privacy-preserving conveyance methods" -- what does
"privacy-preserving" refer to?
* Added reference to "Cryptographic Blinding" section in "Security
Considerations" section.

# S 2.2

* [x] Please remove the confusing parenthetical "(typically via layered
attestation)", or place it more appropriately.
  * Removed it.

# S 4

* [x] The term "implicit authentication" referring to digital signatures
sounds a bit odd
  * Removed "implicit".
* [x] It says: "(see [I-D.ietf-rats-uccs])" -- which section?
  * Added "see Sections 3 and 4 of".

# Figure 2

* [x] Unsure why this needed to be changed, but noting that since you
were at it, you could’ve moved the Verifier to the left of the Attester
to avoid crossing the RP, which is visually confusing.
* The thing is, in all the other diagrams, the Attester is on the left
and the Verifier is on the right. We wanted to represent it in the same
way here.
* To make the diagram clearer, represented the crossings as *aasvg*
"wire jumps/hops."

# S 7.2.1

* [x] Unsure how “Synchronization Checks” would work in practice.
  * Added example showing how TUDA implements this.
  * Does this clarify the comment?

# S 11.1

* [x] Why is RFC9683 normative?
  * Good point. Thanks. Made it informative.