Integrated @gmandyam's comments/review.
* [x] **a)** Sec. 7.3.2.2.: Is there a requirement for the handle to be
Verifier-specific with a vednor-defined freshness policy? If so, how do
both the Attester and Verifier ensure that the PubSub Server is
enforcing per-Verifier compliance with the Verifier freshness policy
(e.g. PubSub server communicates handle as part of an attestation
claim)?
* *Henk's reply:* No, but there can be specifications that come with
that prescription.
  * ➡️ Added clarifying sentence at the end of Section 7.3.2.2.

* [x] **b)** Sec. 7.3.2.3: If the Attestation Result can only be
communicated via publication to the PubSub server, then why does the
figure not show the Attestation Result terminating there?
* *Henk’s Response (rephrased, simplified):* Good point — the diagram
already reflects the correct model (the result is published to the
PubSub server), but the *text may be misleading*.
* ➡️ Added clarification by improving/extending the last sentence of
Section 7.3.2.3.

* [x] **c)** Sec. 7.3.2.4: Call flow in figure seems to be overly
limiting. Attester can also subscribe to attestationResult.
* *Henk’s reply (rephrased, simplified):* Agree, but we want to keep
the diagram simple. Maybe this can be described as a “re-application” of
the interaction model.
  * ➡️ Added clarifying sentence at the end of Section 7.3.2.4.

* [x] **d)** Section 9: Need a discussion on security/privacy
considerations for PubSub server. This is different that what is
described in Section 9.3. For instance, if the PubSub server must attest
to the Handle used as part of evidence then this could be discussed in
Section 9. Moreover, "Strong Isolation of Topics" does not directly
address privacy preservation (i.e. limiting dissemination of identifying
information to subscribers that don't require it).
Moreover for "Trusted Association Verification" it is not clear how
this can be achieved in the PubSub server. For instance, are
endorsements specific to the "the trust association between senders and
receivers" required?

  * [x] **d.1)** What does “attest to the Handle” mean?
* Clarification: This seems to ask whether the PubSub server must
vouch for the Handle (e.g., by proving it delivered it, or binding it
cryptographically).
    * ➡️ Added clarifying sentence at the end of Section 9.3.

* [x] **d.2)** Strong Isolation of Topics" does not address privacy
preservation
    * This is a good point: topic isolation != privacy preservation.
* ➡️ Added clarifying sentence for *Strong Isolation of Topics*
bullet point of Section 9.3.

  * [x] **d.3)** "Trusted Association Verification" lacks details
* "Are endorsements required? How are trust relationships
established?": This is currently vague.
* ➡️ Added clarifying sentence for *Trusted Association
Verification* bullet point of Section 9.3.

Signed-off-by: Michael Eckel <michael.eckel@sit.fraunhofer.de>